requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.8.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
